+++ Runtime/Api/Inventory.cs
@@ -5,8 +5,7 @@
 using AccelByte.Core;
 using AccelByte.Models;
 using AccelByte.Api.Interface;
-using System;
-using NUnit.Framework;
+using UnityEngine.Assertions;
 
 namespace AccelByte.Api
 {